가로세로 비율 동일하면 랜덤으로 내려가도록

+++ src/main/java/dev/practice/ad/domain/api/AdRequestCommand.java
@@ -16,8 +16,8 @@ public class AdRequestCommand {
     private Integer duplicatedNum;
     private AdsType adsType;
     private String unit;
-    private Integer width;
-    private Integer height;
+    private Float width;
+    private Float height;
     private String userIp;
     private String sdkVersion;
